+++ lib/rekindle/web/builder.ex
@@ -247,7 +247,16 @@ defmodule Rekindle.Web.Builder do
   end
 
   defp cargo_options(options),
-    do: Keyword.take(options, [:cargo, :rustc, :timeout, :output_limit, :env])
+    do:
+      Keyword.take(options, [
+        :cargo,
+        :rustc,
+        :rustup,
+        :timeout,
+        :output_limit,
+        :env,
+        :process_env
+      ])
 
   defp toolchain_options(options), do: Keyword.take(options, [:timeout, :env])